# refactor(cubestore): Remove uneeded spawn_blocking (improve perf)