--- a/rust/cubestore/cubestore/Cargo.toml
+++ b/rust/cubestore/cubestore/Cargo.toml
@@ -32,7 +32,7 @@ datafusion = { git = "https://github.com/cube-js/arrow-datafusion", branch = "cu
 csv = "1.1.3"
 bytes = "0.5.4"
 serde_json = "1.0.56"
-futures = "0.3.5"
+futures = "0.3.26"
 smallvec = "1.8.0"
 flexbuffers = "0.1.1"
 byteorder = "1.3.4"
@@ -56,7 +56,7 @@ bigdecimal = { version = "0.2.0", features = ["serde"] }
 rust-s3 = "0.26.3"
 aws-creds = "0.24.1"
 aws-region = "0.22.1"
-deadqueue = "0.1.0"
+deadqueue = "0.2.4"
 reqwest = { version = "0.11.0", features = ["json", "rustls-tls"], default-features = false }
 nanoid = "0.3.0"
 rand = "0.8.0"