Hello!

### Before

```sh
in_process::cache_set_get
                        time:   [10.859 µs 11.142 µs 11.421 µs]
                        change: [-0.6686% +1.4323% +3.6888%] (p = 0.20 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
```

### After

```sh
in_process::cache_set_get
                        time:   [8.2918 µs 8.4782 µs 8.6928 µs]
                        change: [-25.218% -23.466% -21.578%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
```

Thanks